// ui/src/generated/DeploymentDiskInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiskEntry } from "./DiskEntry";
import type { ManagedWorktreeDiskEntry } from "./ManagedWorktreeDiskEntry";

export type DeploymentDiskInfo = { disk: Array<DiskEntry>, managed_worktrees: Array<ManagedWorktreeDiskEntry>, sampled_at: string, };